refactor: consolidate MySQL domain tests
